Add flag to toggle building executable
Some distributors need to use the old Cabal-v1 build systems, which
will build the 'extensions' executable even if it isn't needed (for
instance when this package is only needed as a dependency for stan).

This adds a new 'executable' flag (enabled by default) which will allow
for disabling the 'extensions' executable if it isn't needed.

Signed-off-by: hololeap <hololeap@users.noreply.github.com>